refactor: Web 로또 경로 정리
