+++ src/js/util/console/Trimmer.js
