+++ src/js/util/web/querySelector.js
